Add script to enable and disable the maintenance page in a given namespace
There are some caveats for this script:

This only works on Macs. Sorry Windows. Actually, not sorry. Just being
polite. This is because of how sed works.

action, namespace and hostname_prefix needs to be provided as arguments
to the script or it will not run correctly.

This is a temporary measure to get this working in the shortest space of
time. The template is a direct copy from that of the one in the Editor
that is used whenever a form is published to a namespace. If that ever
changes then this template becomes out of date.

The better implementation would be to put this into the admin dashboard.
Future selves I guess.